requirements.txt

Django==4.1.6
markdown==3.4.1

psycopg2==2.9.4
coverage==7.1.0

lxml==4.9.2

cssselect==1.2.0
fuzzywuzzy==0.18.0

chardet==5.1.0
idna==3.4

certifi==2022.12.7

urllib3==1.26.12

requests==2.28.2
pep8==1.7.1
pyflakes==2.5.0
mccabe==0.7.0
pycodestyle==2.9.1

flake8==5.0.4

configparser==5.3.0

gunicorn==20.1.0
six==1.16.0
boto3==1.26.64
python-dateutil==2.8.2
sqlparse==0.4.3
raven==6.10.0

pytz==2022.7

itsdangerous==2.1.2

django-appconf==1.0.5
django_compressor==4.0
django-bootstrap-form==3.4

django-storages==1.13.2

django-cacheds3storage==0.2.2
django-smoketest==1.2.0

django-debug-toolbar==3.8.1

django-waffle==3.0.0

asgiref==3.5.2

django-extensions==3.2.1
thraxilsettings==5.0.0
